[Issue #6] - Complete install.sh --verify
In refactoring install.sh in preparation for an in-place update,
I have added a --verify action that isolates installation verification
from either Install or Update actions.  In addition to testing, this
offers the user the ability to verify their install and fix any
problems without having to worry about the side effects of performing
an installation or update.

The installation, verification, and update actions are abstracted into
their own functions.

The report_installed_comps function is new and reports on which
components are installed from the perspective of verifying the
installation.  This means that missing components are considered
an error, and all components present is success.