[Issue #60] - Autocomplete of subdirectories working
When a user wants to navigate to a parked directory, they can use
autocomplete to find subdirectories in the target directory by
typing <tab> and <tab><tab>.  With no match prefix, the user can
type <tab><tab> and get a list of every subdirectory in the target.
The user can narrow the search by providing beginning match characters.
If a single match is found, the match will be appended to the ref name
with a trailing "/" so that the user can continue to use autocomplete
to follow the directory tree down.

Add a To Do list for version 2.0.0 to keep track of a few items while
I'm working offline.