[Issue #60] - Autocomplete ref names for del & expand
Autocomplete takes into consideration if the previous word is an
option (beginning with -).  If performing delete or expand,
autocomplete ref names.  Otherwise, do not perform any autocomplete.
It doesn't make sense, for example, to peform autocomplete for add
because the user needs to pick a unique name; there aren't any
known options to provide the user.